move sdk.ts
