+++ packages/shared/src/types/sdk.ts
@@ -1,5 +1,5 @@
 import type { Flowglad as FlowgladNode } from '@flowglad/node'
-import type { Price } from './types/catalog'
+import type { Price } from './catalog'
 
 export enum FlowgladActionKey {
   GetCustomerBilling = 'customers/billing',
